+++ NEWS.md
@@ -12,7 +12,8 @@ Initial features
 * `set_bookmarks()` sets (replaces) pdf bookmarks information (#2).
   `set_bookmarks()` will try to use the following helper functions in the following order:
 
-  1. `set_bookmarks_pdftk()` which wraps `pdftk` command-line tool
+  1. `set_bookmarks_gs()` which wraps `ghostscript` command-line tool
+  2. `set_bookmarks_pdftk()` which wraps `pdftk` command-line tool
 
 * `docinfo()` provides an `{R6}` object to represent pdf documentation info entries.
 * `get_docinfo()` reads in pdf documentation info entries from a pdf file.